make int to be int32 for RPC

+++ delayed_task_scheduler.hxx
@@ -6,7 +6,7 @@ namespace cornerstone {
     __interface_body__(delayed_task_scheduler)
 
     public:
-        virtual void schedule(std::shared_ptr<delayed_task>& task, long milliseconds);
+        virtual void schedule(std::shared_ptr<delayed_task>& task, int32 milliseconds);
     };
 }
